Move exceptions to Utility project
* Add new Utlity project
* Move ItemNotFoundException to Utility project
* Refactor code
* Fix test cases
